Implement multiplayer demo
