+++ client/index.html
@@ -37,7 +37,6 @@
   <script type="text/javascript" src="scripts/services/firebase.js"></script>
   <script type="text/javascript" src="scripts/services/comm.js"></script>
   <script type="text/javascript" src="scripts/services/panel.js"></script>
-  <script type="text/javascript" src="scripts/services/puzzle.js"></script>
   <script type="text/javascript" src="scripts/services/angular-knob.js"></script>
   <script type="text/javascript" src="lib/angular-ui-slider/src/slider.js"></script>
   <!-- controllers -->
