+++ client/scripts/app.js
@@ -4,7 +4,6 @@ angular.module('app', [
   'ui.router',
   'app.comm',
   'app.panel',
-  'app.puzzle',
   'app.intro',
   'app.gameFrame',
   'app.game',
