+++ package.json
@@ -20,7 +20,8 @@
   "homepage": "https://github.com/team-jems/hrgame#readme",
   "dependencies": {
     "bower": "^1.6.2",
-    "express": "^4.13.3"
+    "express": "^4.13.3",
+    "firebase": "^2.3.1"
   },
   "engines": {
     "node": "0.12.7"
